Enhance PacketRegistry and hosting workflows
- Added `.nuget/` to `.gitignore` to exclude NuGet files.
- Updated `PacketRegistry` docs with new APIs and workflows.
- Introduced `PacketRegister` as a static convenience API.
- Added namespace and assembly-path-based packet registration.
- Enhanced `NetworkApplication` builder with new methods.
- Refactored and expanded tests for packet discovery methods.
- Added new packet classes for testing serialization/deserialization.
- Updated `restore.ps1` for sequential restores and better logging.
